fix: ajuste cliente service post cliente

+++ backend/client-service/src/main/java/com/example/client_service/dto/AdicionarMilhasDTO.java
@@ -4,13 +4,13 @@ import java.math.BigDecimal;
 
 
 public class AdicionarMilhasDTO {
-    private BigDecimal quantidade;
+    private Integer quantidade;
 
-    public BigDecimal getQuantidade() {
+    public Integer getQuantidade() {
         return quantidade;
     }
 
-    public void setQuantidade(BigDecimal quantidade) {
+    public void setQuantidade(Integer quantidade) {
         this.quantidade = quantidade;
     }
 }
